Share more events between shamans
