+++ src/roles/crazed_shaman.py
@@ -13,64 +13,19 @@ from src.dispatcher import MessageDispatcher
 from src.messages import messages
 from src.events import Event
 
-from src.roles._shaman_helper import havetotem, _get_target, _give_totem, _apply_totem
+from src.roles._shaman_helper import setup_variables, get_totem_target, give_totem
 
-TOTEMS = UserDict()         # type: Dict[users.User, str]
-LASTGIVEN = UserDict()      # type: Dict[users.User, users.User]
-SHAMANS = UserDict()        # type: Dict[users.User, List[users.User]]
+TOTEMS, LASTGIVEN, SHAMANS = setup_variables("crazed shaman", knows_totem=False)
 
 @command("give", "totem", chan=False, pm=True, playing=True, silenced=True, phases=("night",), roles=("crazed shaman",))
-def crazed_shaman_totem(var, wrapper, message, prefix="You"):
+def crazed_shaman_totem(var, wrapper, message):
     """Give a random totem to a player."""
 
-    target = _get_target(var, wrapper, message, LASTGIVEN)
+    target = get_totem_target(var, wrapper, message, LASTGIVEN)
     if not target:
         return
 
-    _give_totem(var, wrapper, target, prefix, set(), "crazed shaman", "", SHAMANS)
-
-@event_listener("del_player")
-def on_del_player(evt, var, user, mainrole, allroles, death_triggers):
-    for a,(b,c) in list(SHAMANS.items()):
-        if user in (a, b, c):
-            del SHAMANS[a]
-
-@event_listener("night_acted")
-def on_acted(evt, var, user, actor):
-    if user in SHAMANS:
-        evt.data["acted"] = True
-
-@event_listener("get_special")
-def on_get_special(evt, var):
-    evt.data["special"].update(get_players(("crazed shaman",)))
-
-@event_listener("exchange_roles")
-def on_exchange(evt, var, actor, target, actor_role, target_role):
-    actor_totem = None
-    target_totem = None
-    if actor_role == "crazed shaman":
-        actor_totem = TOTEMS.pop(actor)
-        if actor in SHAMANS:
-            del SHAMANS[actor]
-        if actor in LASTGIVEN:
-            del LASTGIVEN[actor]
-
-    if target_role == "crazed shaman":
-        target_totem = TOTEMS.pop(target)
-        if target in SHAMANS:
-            del SHAMANS[target]
-        if target in LASTGIVEN:
-            del LASTGIVEN[target]
-
-    if target_totem:
-        TOTEMS[actor] = target_totem
-    if actor_totem:
-        TOTEMS[target] = actor_totem
-
-@event_listener("chk_nightdone")
-def on_chk_nightdone(evt, var):
-    evt.data["actedcount"] += len(SHAMANS)
-    evt.data["nightroles"].extend(get_players(("crazed shaman",)))
+    SHAMANS[wrapper.source] = give_totem(var, wrapper, target, prefix="You", tags=set(), role="crazed shaman", msg="")
 
 @event_listener("player_win")
 def on_player_win(evt, var, user, role, winner, survived):
@@ -94,22 +49,11 @@ def on_transition_day_begin(evt, var):
                 target = random.choice(ps)
                 dispatcher = MessageDispatcher(shaman, shaman)
 
-                _give_totem(var, dispatcher, target, messages["random_totem_prefix"], set(), "crazed shaman", "", SHAMANS)
+                SHAMANS[shaman] = give_totem(var, dispatcher, target, prefix=messages["random_totem_prefix"], tags=set(), role="crazed shaman", msg="")
             else:
                 LASTGIVEN[shaman] = None
         elif shaman not in SHAMANS:
             LASTGIVEN[shaman] = None
-
-@event_listener("transition_day_begin", priority=7)
-def on_transition_day_begin2(evt, var):
-    for shaman, (victim, target) in SHAMANS.items():
-        _apply_totem(TOTEMS[shaman], shaman, victim)
-
-        if target is not victim:
-            shaman.send(messages["totem_retarget"].format(victim))
-        LASTGIVEN[shaman] = victim
-
-    havetotem.extend(sorted(filter(None, LASTGIVEN.values())))
 
 @event_listener("transition_night_end", priority=2.01)
 def on_transition_night_end(evt, var):
@@ -144,37 +88,10 @@ def on_transition_night_end(evt, var):
             shaman.send(messages["shaman_notify"].format("crazed shaman", "random "))
         shaman.send("Players: " + ", ".join(p.nick for p in pl))
 
-@event_listener("begin_day")
-def on_begin_day(evt, var):
-    SHAMANS.clear()
-
-@event_listener("reset")
-def on_reset(evt, var):
-    TOTEMS.clear()
-    LASTGIVEN.clear()
-    SHAMANS.clear()
-
 @event_listener("succubus_visit")
 def on_succubus_visit(evt, var, succubus, target):
     if target in SHAMANS and SHAMANS[target][1] in get_all_players(("succubus",)):
         target.send(messages["retract_totem_succubus"].format(SHAMANS[target][1]))
         del SHAMANS[target]
-
-@event_listener("revealroles_role")
-def on_revealroles(evt, var, wrapper, user, role):
-    if user in TOTEMS:
-        if user in SHAMANS:
-            evt.data["special_case"].append("giving {0} totem to {1}".format(TOTEMS[user], SHAMANS[user][0]))
-        elif var.PHASE == "night":
-            evt.data["special_case"].append("has {0} totem".format(TOTEMS[user]))
-        elif user in LASTGIVEN and LASTGIVEN[user]:
-            evt.data["special_case"].append("gave {0} totem to {1}".format(TOTEMS[user], LASTGIVEN[user]))
-
-@event_listener("get_role_metadata")
-def on_get_role_metadata(evt, var, kind):
-    if kind == "night_kills":
-        # only add shamans here if they were given a death totem
-        # even though retribution kills, it is given a special kill message
-        evt.data["crazed shaman"] = list(TOTEMS.values()).count("death")
 
 # vim: set sw=4 expandtab:
